refactor: change project structure
