+++ src/app/auth/components/admin-panel/admin-panel.component.ts
