+++ src/app/auth/components/login/login.component.html
@@ -1,0 +1,13 @@
+<div class="login-container">
+  <form (submit)="handleLogin()" class="login-form">
+    <div class="input-container">
+      <label for="login-username">Username: *</label>
+      <input type="text" id="login-username" name="username">
+    </div>
+    <div class="input-container">
+      <label for="login-password">Password: *</label>
+      <input type="password" id="login-password" name="password">
+    </div>
+    <button mat-raised-button color="primary" type="submit">Log in</button>
+  </form>
+</div>
