+++ src/app/auth/components/login/login.component.scss
@@ -1,0 +1,11 @@
+.login-form{
+  display:flex;
+  flex-direction:column;
+  justify-content:space-evenly;
+  align-items:center;
+  gap:20px;
+}
+.input-container{
+  display:flex;
+  flex-direction:column;
+}
