+++ src/app/auth/components/registration/registration.component.spec.ts
