+++ src/app/auth/components/registration/registration.component.ts
