+++ src/app/core/components/header/header.component.scss
@@ -37,8 +37,6 @@ header{
   }
 
 }
-
-
 .play-icon-container{
   display:flex;
   justify-content:center;
