+++ src/app/core/pages/page-not-found/page-not-found.component.ts
@@ -1,12 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-page-not-found',
   templateUrl: './page-not-found.component.html',
   styleUrls: ['./page-not-found.component.scss'],
 })
-class PageNotFoundComponent implements OnInit {
-  ngOnInit(): void {
-  }
+class PageNotFoundComponent {
 }
 export default PageNotFoundComponent;
