--- src/app/login/login.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-login',
-  templateUrl: './login.component.html',
-  styleUrls: ['./login.component.scss'],
-})
-class LoginComponent implements OnInit {
-  ngOnInit(): void {
-  }
-}
-export default LoginComponent;
