--- src/app/page-not-found/page-not-found.component.html
@@ -1,1 +1,0 @@
-<p>page-not-found works!</p>
