+++ src/app/youtube/components/card/card.component.html
@@ -16,5 +16,5 @@
     </div>
   </div>
   <h2 class="card-title">{{cardInfo.snippet.title}}</h2>
-  <button mat-raised-button class="card-button" color="primary">More...</button>
+  <button mat-raised-button class="card-button" color="primary" [routerLink]="cardInfo.id">More...</button>
 </mat-card>
