+++ src/app/youtube/components/card/card.component.scss
