+++ src/app/youtube/components/card/card.component.ts
@@ -1,15 +1,12 @@
-import { Component, Input, OnInit } from '@angular/core';
-import SearchItem from '../search-item.model';
+import { Component, Input } from '@angular/core';
+import SearchItem from '../../models/search-item.model';
 
 @Component({
   selector: 'app-card',
   templateUrl: './card.component.html',
   styleUrls: ['./card.component.scss'],
 })
-class CardComponent implements OnInit {
+class CardComponent {
   @Input() cardInfo! : SearchItem;
-
-  ngOnInit(): void {
-  }
 }
 export default CardComponent;
