+++ src/app/youtube/directives/card.directive.ts
