+++ src/app/youtube/models/search-item.model.ts
