+++ src/app/youtube/pages/search-page/search-page.component.scss
