+++ src/app/youtube/pipes/filter/filter.pipe.ts
@@ -1,5 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import SearchItem from 'src/app/search-page/search-item.model';
+import SearchItem from 'src/app/youtube/models/search-item.model';
 
 @Pipe({
   name: 'filter',
